Add user authorization control and manage listings page

+++ app/Http/Controllers/ListingController.php
@@ -58,6 +58,11 @@ class ListingController extends Controller
 
     // Update Listing data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -79,8 +84,21 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings()->get()
+        ]);
     }
 }
